data-platform/streaming/requirements.txt
apache-beam[gcp]==2.53.0
google-cloud-bigquery==3.25.0
confluent-kafka==2.3.0
python-dateutil>=2.8.0
